whatpm/t/tree-test-1.dat

#data
<!DOCTYPE HTML><html title="t"><head><title></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   title="t"
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html><head title="t"><title></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     title="t"
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html><head><title class="t"></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|       class="t"
|   <body>

#data
<!DOCTYPE HTML><html><head></head><body><title class="a"></title></body></html>
#errors
title in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|       class="a"
|   <body>

#data
<!DOCTYPE HTML><html><head></head><body></body><title class="a"></title></html>
#errors
data after body
title in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|       class="a"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body title="t"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     title="t"

#data
<!DOCTYPE HTML><html><head><title></title><meta name="a"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <meta>
|       name="a"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body><meta name="a"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <meta>
|       name="a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body><meta name="a"></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <meta>
|       name="a"

#data
<!DOCTYPE HTML><html><head><title></title><link rel="a"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <link>
|       rel="a"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body><link rel="a"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <link>
|       rel="a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body><link rel="a"></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <link>
|       rel="a"

#data
<!DOCTYPE HTML><html><head><title></title><base href="http://"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <base>
|       href="http://"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body><base href="http://"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <base>
|       href="http://"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body><base href="http://"></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <base>
|       href="http://"

#data
<!DOCTYPE HTML><p><b><p>X
#errors
<b> is not closed before second <p>
<b> is not closed before end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|     <p>
|       <b>
|         "X"

#data
<!DOCTYPE HTML><b><p>
#errors
<b> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       <p>

#data
<!DOCTYPE HTML><b><p>X
#errors
<b> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       <p>
|         "X"

#data
<!DOCTYPE HTML><ins><p>XX</ins>YYY
#errors
1,32,not closed:p
1,34,not closed:ins
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ins>
|       <p>
|         "XXYYY"

## when script is disabled
#data
<!DOCTYPE HTML><noscript>aaa<noscript>bbb</noscript>ccc</noscript>
#errors
in noscript:#character
unmacthed noscript end tag
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>
|     "aaa"
|     <noscript>
|       "bbb"
|     "ccc"

## when script is disabled
#data
<!DOCTYPE HTML><body><noscript>aaa<noscript>bbb</noscript>ccc</noscript>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noscript>
|       "aaa"
|       <noscript>
|         "bbb"
|       "ccc"

## when script is disabled
#data
<!DOCTYPE HTML><head><noscript><noscript>
#errors
<noscript> in noscript
no </noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>

## when script is disabled
#data
<!DOCTYPE HTML><head><noscript><noscript> 
#errors
<noscript> in noscript
no </noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|       " "

## when script is disabled
#data
<!DOCTYPE HTML><head><noscript> <noscript> 
#errors
<noscript> in noscript
no </noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|       "  "

#data
<!DOCTYPE HTML><iframe>
#errors
no </iframe>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <iframe>

#data
<!DOCTYPE HTML><iframe></iframe>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <iframe>

#data
<!DOCTYPE HTML><iframe>a</iframe>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <iframe>
|       "a"

#data
<!DOCTYPE HTML><noframes>
#errors
no </noframes>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>

#data
<!DOCTYPE HTML><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>

#data
<!DOCTYPE HTML><noframes>a</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>
|       "a"

#data
<!DOCTYPE HTML><noframes>
#errors
no </noframes>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>

#data
<!DOCTYPE HTML><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>

#data
<!DOCTYPE HTML><noframes>a</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>
|       "a"

#data
<!DOCTYPE HTML><xmp>
#errors
no </xmp>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <xmp>

#data
<!DOCTYPE HTML><xmp></xmp>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <xmp>

#data
<!DOCTYPE HTML><xmp>a</xmp>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <xmp>
|       "a"

#data
<!DOCTYPE HTML><html><head><title></title><style></style></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <style>
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title><style type="text/css"></style></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <style>
|       type="text/css"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title><style type="text/css">p {}</style></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <style>
|       type="text/css"
|       "p {}"
|   <body>

#data
<!DOCTYPE HTML><html><style type="text/css"></style><title></title></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <style>
|       type="text/css"
|     <title>
|   <body>

#data
<!DOCTYPE HTML><style type="text/css"></style><title></title><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <style>
|       type="text/css"
|     <title>
|   <body>

#data
<!DOCTYPE HTML><title></title><body><p>a<style type="text/css"></style>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <p>
|       "a"
|       <style>
|         type="text/css"

#data
<!DOCTYPE HTML><form><form></form>
#errors
form in form
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>

#data
<!DOCTYPE HTML><ul><li>a<li>c</ul>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         "a"
|       <li>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><ul><li>a<div>b<li>c</ul>d
#errors
div closed by li
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         "a"
|         <div>
|           "b"
|       <li>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<dt>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<dd>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<dt>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<dd>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<div>b<dt>c</dl>d
#errors
div closed by dt
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|         <div>
|           "b"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<div>b<dd>c</dl>d
#errors
div closed by dd
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|         <div>
|           "b"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<div>b<dt>c</dl>d
#errors
div closed by dt
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|         <div>
|           "b"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<div>b<dd>c</dl>d
#errors
div closed by dd
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|         <div>
|           "b"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><body><form><div>a</form>b
#errors
div not closed before </form>
div not closed before EOF
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <div>
|         "ab"

#data
<!DOCTYPE HTML><ul></body>
#errors
ul not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>

#data
<!DOCTYPE HTML><ul><li></body>
#errors
ul not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>

#data
<!DOCTYPE HTML><dl></body>
#errors
dl not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>

#data
<!DOCTYPE HTML><dl><dt></body>
#errors
dl not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>

#data
<!DOCTYPE HTML><dl><dd></body>
#errors
dl not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>

#data
<!DOCTYPE HTML><p></body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><table><tr><td></body>
#errors
table not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><table><tr><th></body>
#errors
table not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>

#data
<!DOCTYPE HTML><table><thead><tr><td></body>
#errors
table not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><isindex>
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <p>
|         <label>
|           "This is a searchable index. Insert your search keywords here: "
|           <input>
|             name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex action="http://test/">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       action="http://test/"
|       <hr>
|       <p>
|         <label>
|           "This is a searchable index. Insert your search keywords here: "
|           <input>
|             name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex prompt="Input:">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <p>
|         <label>
|           "Input:"
|           <input>
|             name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex name="textarea">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <p>
|         <label>
|           "This is a searchable index. Insert your search keywords here: "
|           <input>
|             name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex type="checkbox">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <p>
|         <label>
|           "This is a searchable index. Insert your search keywords here: "
|           <input>
|             name="isindex"
|             type="checkbox"
|       <hr>

#data
<!DOCTYPE HTML><form><isindex></form>
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>

#data
<!DOCTYPE HTML><h1><a><h2>
#errors
h1 not closed
a not closed
h2 not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <a>
|         <h2>

#data
  <p>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>

#data
  a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     "a"

#data
<!---->  <p>
#errors
no DOCTYPE
#document
| <!--  -->
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML>  <p>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><div></div></div>
#errors
unmatched </div>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <div>

#data
<!DOCTYPE HTML><div><span></div>
#errors
unclosed <span>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <div>
|       <span>

#data
<!DOCTYPE HTML><pre><!---->
a</pre>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <pre>
|       <!--  -->
|       "
a"

#data
<!DOCTYPE HTML><textarea><!---->
a</textarea>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "<!---->
a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html> 
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     " "

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>  
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "  "

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>a
#errors
char after </body>
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html> a
#errors
char after </html>
char after </body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     " a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>a 
#errors
char after </html>
char after </body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "a "

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>aaa
#errors
char after </html>
char after </body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "aaa"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html><!---->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
| <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html><!--aa-->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
| <!-- aa -->

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html><!DOCTYPE HTML>
#errors
<!DOCTYPE> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html><p>
#errors
start tag after </html>
start tag after </body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html></p>
#errors
end tag after </html>
end tag after </body>
</p> interpreted as <p></p>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>  <p>
#errors
start tag after </html>
start tag after </body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "  "
|     <p>

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>aa<p>
#errors
char after </html>
char after </body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "aa"
|     <p>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> 
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>  
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   "  "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a
#errors
char after </html>
char after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> a
#errors
char after </html>
char after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a 
#errors
char after </html>
char after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><!---->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
| <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><!--a-->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
| <!-- a -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> <!---->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "
| <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a<!---->
#errors
char after </html>
char after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><!DOCTYPE HTML>
#errors
<!DOCTYPE> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> <!DOCTYPE HTML>
#errors
<!DOCTYPE> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><p>
#errors
start tag after </html>
start tag after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> <p>
#errors
start tag after </html>
start tag after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html></p>
#errors
end tag after </html>
end tag after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> </p>
#errors
end tag after </html>
end tag after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html></html>
#errors
</html> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a</html>
#errors
char after </html>
char after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes></noframes>
#errors
<noframes> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes>
#errors
<noframes> after </html>
no </noframes>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes><p></noframes>
#errors
<noframes> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>
|     "<p>"

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes></noframes>a
#errors
<noframes> after </html>
char after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> <noframes></noframes> 
#errors
<noframes> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "
|   <noframes>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset> </frameset></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|     " "

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body> </html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     " "

#data
<!DOCTYPE HTML><frameset></frameset><noframes>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><frameset></frameset><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><frameset></frameset><noframes>abcdefg
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "abcdefg"

#data
<!DOCTYPE HTML><frameset></frameset><noframes>abcdefg</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "abcdefg"

#data
<!DOCTYPE HTML><frameset></frameset><noframes>aa<p>bb</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "aa<p>bb"

#data
<!DOCTYPE HTML><frameset></frameset><noframes><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "<noframes>"

#data
<!DOCTYPE HTML><frameset></frameset><noframes></noframe></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "</noframe>"

#data
</body></html><html a>
#errors
no <!DOCTYPE>
<html> after </html>
#document
| <html>
|   a=""
|   <head>
|   <body>

#data
<body></body></html><html a>
#errors
no <!DOCTYPE>
<html> after </html>
<html> is not the first start tag
#document
| <html>
|   a=""
|   <head>
|   <body>

#data
<frameset></frameset></html><html a>
#errors
no <!DOCTYPE>
<html> after </html>
<html> is not the first start tag
#document
| <html>
|   a=""
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><html><head></head><frameset></frameset></html><html a>
#errors
<html> after </html>
<html> is not the first start tag
#document
| <!DOCTYPE HTML>
| <html>
|   a=""
|   <head>
|   <frameset>

#data
<frameset></frameset>
#errors
no <!DOCTYPE>
#document
| <html>
|   <head>
|   <frameset>

#data
<frameset></frameset></frameset>
#errors
no <!DOCTYPE>
unmatched </frameset>
#document
| <html>
|   <head>
|   <frameset>

#data
<frameset></frameset></frameset></frameset>
#errors
no <!DOCTYPE>
unmatched </frameset>
unmatched </frameset>
#document
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><table><caption>test</caption>a</table>
#errors
a in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <table>
|       <caption>
|         "test"

#data
<!DOCTYPE HTML><table><caption>test</table>a
#errors
no </caption>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "test"
|     "a"

#data
<!DOCTYPE HTML><table><caption>test</body>a
#errors
unmatched </body>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</col>a
#errors
unmatched </col>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</colgroup>a
#errors
unmatched </colgroup>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</html>a
#errors
unmatched </html>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</html> 
#errors
unmatched </html>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "test "

#data
<!DOCTYPE HTML><table><caption>test</tbody>a
#errors
unmatched </tbody>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</thead>a
#errors
unmatched </thead>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</tfoot>a
#errors
unmatched </tfoot>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</tr>a
#errors
unmatched </tr>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</td>a
#errors
unmatched </td>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</th>a
#errors
unmatched </th>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>a<caption>b
#errors
no </caption>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "a"
|       <caption>
|         "b"

#data
<!DOCTYPE HTML><table><caption>a<col>b
#errors
no </caption>
"b" in table
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <colgroup>
|         <col>

#data
<!DOCTYPE HTML><table><caption>a<colgroup>b
#errors
no </caption>
"b" in colgroup
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <colgroup>

#data
<!DOCTYPE HTML><table><caption>a<thead>b
#errors
no </caption>
"b" in thead
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <thead>

#data
<!DOCTYPE HTML><table><caption>a<tbody>b
#errors
no </caption>
"b" in tbody
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <tbody>

#data
<!DOCTYPE HTML><table><caption>a<tfoot>b
#errors
no </caption>
"b" in tfoot
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <tfoot>

#data
<!DOCTYPE HTML><table><caption>a<tr>b
#errors
no </caption>
"b" in tr
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><caption>a<td>b
#errors
no </caption>
no <tr>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "a"
|       <tbody>
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><caption>a<th>b
#errors
no </caption>
no <tr>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "a"
|       <tbody>
|         <tr>
|           <th>
|             "b"

#data
<!DOCTYPE HTML>a<caption>b
#errors
<caption> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<col>b
#errors
<col> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<colgroup>b
#errors
<colgroup> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<frame>b
#errors
<frame> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<frameset>b
#errors
<frameset> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<frameset><frame>b
#errors
<frameset> in body
<frame> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<head>b
#errors
<head> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<head><link>b
#errors
<head> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <link>
|     "b"

#data
<!DOCTYPE HTML>a<option>b
#errors
<option> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<optgroup>b
#errors
<optgroup> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<tbody>b
#errors
<tbody> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<tbody><tr>b
#errors
<tbody> in body
<tr> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<thead>b
#errors
<thead> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<tfoot>b
#errors
<tfoot> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<tr>b
#errors
<tr> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<td>b
#errors
<td> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<th>b
#errors
<th> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML><table><tr><td>a<tr><td>b</table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><tr><td>a<thead><tr><td>b</table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|       <thead>
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><tr><td>a<tbody><tr><td>b</table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|       <tbody>
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><tr><td>a<tfoot><tr><td>b</table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|       <tfoot>
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><tr><td>a<caption>b</caption></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|       <caption>
|         "b"

#data
<!DOCTYPE HTML><table><tr><td>a</th></table>
#errors
unmatched </th>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"

#data
<!DOCTYPE HTML><table><tr><td>a</th>b</table>
#errors
unmatched </th>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</body>b
#errors
unmatched </body>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</html>b
#errors
unmatched </html>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</caption>b
#errors
unmatched </caption>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</colgroup>b
#errors
unmatched </colgroup>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</col>b
#errors
unmatched </col>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</thead>b
#errors
unmatched </thead>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</tfoot>b
#errors
unmatched </tfoot>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tfoot><tr><td>a</tbody>b
#errors
unmatched </tbody>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</tbody>b
#errors
"b" in table
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"

#data
<!DOCTYPE HTML><table><thead><tr><td>a</thead>b
#errors
"b" in table
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <thead>
|         <tr>
|           <td>
|             "a"

#data
<!DOCTYPE HTML><table><tfoot><tr><td>a</tfoot>b
#errors
"b" in table
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <tfoot>
|         <tr>
|           <td>
|             "a"

#data
<!DOCTYPE HTML><table><tbody><caption>a</caption>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <caption>
|         "a"

#data
<!DOCTYPE HTML><table><tbody><thead><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody><tfoot><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <tfoot>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody><tbody><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><thead><tfoot><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|       <tfoot>
|         <tr>

#data
<!DOCTYPE HTML><table><thead><tbody><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><thead><thead><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML><table a><thead><table b>
#errors
<table> in table row group
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <thead>
|     <table>
|       b=""

#data
<!DOCTYPE HTML><table a><tbody><table b>
#errors
<table> in table row group
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <tbody>
|     <table>
|       b=""

#data
<!DOCTYPE HTML><table a><tfoot><table b>
#errors
<table> in table row group
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <tfoot>
|     <table>
|       b=""

#data
<!DOCTYPE HTML><table a><tr><table b>
#errors
<table> in table row
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <tbody>
|         <tr>
|     <table>
|       b=""

#data
<!DOCTYPE HTML><table><thead></tbody><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML><table><thead></thead><tr>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><thead></tfoot><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></thead><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></tbody><tr>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></tfoot><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tfoot></thead><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <tr>

#data
<!DOCTYPE HTML><table><tfoot></tbody><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <tr>

#data
<!DOCTYPE HTML><table><tfoot></tfoot><tr>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></tr><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></body><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></html><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></td><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></th><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></caption><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></col><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></colgroup><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
